atualiza bugs no projeto 13 (FINANCE TRACKER) e inclui CRUD completo

+++ ex13-gerenciador_financas/financetracker.cs
@@ -26,7 +26,7 @@ namespace FinanceTracker
         MostrarMenu();
         bool validMenuChoice = int.TryParse(Console.ReadLine(), out int choice);
 
-        while (!validMenuChoice || choice < 1 || choice > 7)
+        while (!validMenuChoice || choice < 1 || choice > 8)
         {
           System.Console.WriteLine("incorreto. insira novamente: ");
           validMenuChoice = int.TryParse(Console.ReadLine(), out choice);
@@ -36,23 +36,26 @@ namespace FinanceTracker
           case 1:
             AdicionarCategoria(connectionString);
             break;
-            
+
           case 2:
             ListarCategorias(connectionString);
             break;
           case 3:
-            AdicionarTransacao(connectionString);
+            DeletarCategoria(connectionString);
             break;
           case 4:
-            ListarTransacoes(connectionString);
+            AdicionarTransacao(connectionString);
             break;
           case 5:
-            DeletarTransacao(connectionString);
+            ListarTransacoes(connectionString);
             break;
           case 6:
-            VerSaldoAtual(connectionString);
+            DeletarTransacao(connectionString);
             break;
           case 7:
+            VerSaldoAtual(connectionString);
+            break;
+          case 8:
             running = false;
             break;
           default:
@@ -88,18 +91,21 @@ namespace FinanceTracker
     }
     static void MostrarMenu()
     {
-      Console.WriteLine("\nmenu finance tracker:");
-      Console.WriteLine("1. adicionar categoria");
-      Console.WriteLine("2. listar categorias\n");
-      Console.WriteLine("3. adicionar transação");
-      Console.WriteLine("4. listar transações");
-      Console.WriteLine("5. deletar transação \n");
-      Console.WriteLine("6. ver saldo atual");
-      Console.WriteLine("7. sair");
+      Console.Clear();
+      Console.WriteLine("\nmenu do app: FINANCE TRACKER. uma solução prática para suas finanças pessoais.:");
+      Console.WriteLine("1. adicionar categoria de renda");
+      Console.WriteLine("2. listar categorias");
+      Console.WriteLine("3. deletar categoria\n\n");
+      Console.WriteLine("4. adicionar transação em uma categoria de renda");
+      Console.WriteLine("5. listar transações");
+      Console.WriteLine("6. deletar transação \n");
+      Console.WriteLine("7. ver saldo atual");
+      Console.WriteLine("8. sair");
     }
 
     static void AdicionarCategoria(string connectionString)
     {
+      Console.Clear();
       Console.WriteLine("\nadicionar categoria");
 
       Console.WriteLine("insira o tipo da categoria que quer adicionar (1 para entrada e 2 para saída): ");
@@ -132,10 +138,67 @@ namespace FinanceTracker
       command.Parameters.AddWithValue("$cType", sqlCategoryType);
 
       command.ExecuteNonQuery();
+      Console.ReadKey();
+    }
+
+    static void DeletarCategoria(string connectionString)
+    {
+      Console.Clear();
+      Console.WriteLine("\nlistar categorias\n");
+      using var connCategory = new SqliteConnection(connectionString);
+      connCategory.Open();
+      using var command = connCategory.CreateCommand();
+      command.CommandText = "SELECT * FROM categories";
+      using var reader = command.ExecuteReader();
+
+      int quantidadeCategorias = 0;
+      var idsValidosCategorias = new List<int>();
+
+      while (reader.Read())
+      {
+        int catId = reader.GetInt32(0);
+        string catName = reader.GetString(1);
+        string catType = reader.GetString(2);
+
+        Console.WriteLine($"id da categoria: {catId}\nnome da categoria: {catName}\ntipo da categoria: {catType}\n\n");
+
+        idsValidosCategorias.Add(catId);
+        quantidadeCategorias++;
+      }
+
+      reader.Close();
+
+      if (quantidadeCategorias < 1)
+      {
+        System.Console.WriteLine("não existem categorias para excluir.");
+        return;
+      }
+
+      System.Console.WriteLine("insira o id da categoria que quer excluir (0 para sair): ");
+      bool validTransactionDeleteId = int.TryParse(Console.ReadLine(), out int tIdCategoryDelete);
+
+      if (tIdCategoryDelete == 0) return;
+      
+      while (!validTransactionDeleteId || !idsValidosCategorias.Contains(tIdCategoryDelete))
+      {
+        System.Console.WriteLine("id da categoria incorreto. insira novamente: ");
+        validTransactionDeleteId = int.TryParse(Console.ReadLine(), out tIdCategoryDelete);
+
+        if (tIdCategoryDelete == 0) return;
+      }
+
+      command.CommandText = "DELETE FROM transactions WHERE category_id = $catId";
+      command.Parameters.AddWithValue("$catId", tIdCategoryDelete);
+      command.ExecuteNonQuery();
+
+      command.CommandText = "DELETE FROM categories WHERE id = $catId";
+      command.ExecuteNonQuery();
+      Console.ReadKey();
     }
 
     static void ListarCategorias(string connectionString)
     {
+      Console.Clear();
       Console.WriteLine("\nlistar categorias\n");
       using var connListar = new SqliteConnection(connectionString);
       connListar.Open();
@@ -151,10 +214,12 @@ namespace FinanceTracker
 
         Console.WriteLine($"id da categoria: {catId}\nnome da categoria: {catName}\ntipo da categoria: {catType}\n\n");
       }
+      Console.ReadKey();
     }
 
     static void AdicionarTransacao(string connectionString)
     {
+      Console.Clear();
       Console.WriteLine("\nadicionar transação");
       using var connTransacao = new SqliteConnection(connectionString);
       connTransacao.Open();
@@ -162,7 +227,9 @@ namespace FinanceTracker
       command.CommandText = "SELECT * FROM categories";
       using var reader = command.ExecuteReader();
 
+      var idsValidosCategorias = new List<int>();
       int quantidadeCategorias = 0;
+
       Console.WriteLine("categorias disponíveis: \n");
       while (reader.Read())
       {
@@ -171,18 +238,29 @@ namespace FinanceTracker
         string catType = reader.GetString(2);
 
         Console.WriteLine($"id da categoria: {catId}\nnome da categoria: {catName}\ntipo da categoria: {catType}\n\n");
+        idsValidosCategorias.Add(catId);
         quantidadeCategorias++;
       }
 
       reader.Close();
 
-      System.Console.WriteLine("insira o id da categoria que quer inserir uma transação: ");
+      if (quantidadeCategorias == 0)
+      {
+        System.Console.WriteLine("não há categorias disponíveis para adicionar uma nova transação...");
+        Console.ReadKey();
+        return;
+      }
+      System.Console.WriteLine("insira o id da categoria que quer inserir uma transação (0 para sair): ");
       bool validId = int.TryParse(Console.ReadLine(), out int idCategoria);
 
-      while (!validId || idCategoria < 1 || idCategoria > quantidadeCategorias)
+      if (idCategoria == 0) return;
+
+      while (!validId || !idsValidosCategorias.Contains(idCategoria))
       {
         System.Console.WriteLine("inválido ou id de categoria não existente. insira novamente: ");
         validId = int.TryParse(Console.ReadLine(), out idCategoria);
+
+        if (idCategoria == 0) return;
       }
 
       System.Console.WriteLine("insira a descrição da sua transação: ");
@@ -208,10 +286,12 @@ namespace FinanceTracker
       command.Parameters.AddWithValue("$catId", idCategoria);
 
       command.ExecuteNonQuery();
+      Console.ReadKey();
     }
 
     static void ListarTransacoes(string connectionString)
     {
+      Console.Clear();
       Console.WriteLine("\nlistar transações\n");
       using var connListTransacao = new SqliteConnection(connectionString);
       connListTransacao.Open();
@@ -229,11 +309,13 @@ namespace FinanceTracker
         string cName = reader.GetString(4);
         string cType = reader.GetString(5);
 
-        Console.WriteLine($"CATEGORIA: {cName}\nTIPO DA CATEGORIA: {cType}\nid da transação: {tId}\ndescrição da transação: {tDescription}\nvalor da transação: {tAmount}\ndata da transação: {tDate}\n\n");
+        Console.WriteLine($"CATEGORIA: {cName}\nTIPO DA CATEGORIA: {cType}\nid da transação: {tId}\ndescrição da transação: {tDescription}\nvalor da transação: R${tAmount:F2}\ndata da transação: {tDate}\n\n");
       }
+      Console.ReadKey();
     }
     static void DeletarTransacao(string connectionString)
     {
+      Console.Clear();
       Console.WriteLine("\ndeletar transação\n");
       using var connListTransacaoDeleteTransacao = new SqliteConnection(connectionString);
       connListTransacaoDeleteTransacao.Open();
@@ -243,6 +325,8 @@ namespace FinanceTracker
       using var reader = command.ExecuteReader();
 
       int quantidadeTransacoes = 0;
+      var idsValidos = new List<int>();
+
       while (reader.Read())
       {
         int tId = reader.GetInt32(0);
@@ -252,32 +336,41 @@ namespace FinanceTracker
         string cName = reader.GetString(4);
         string cType = reader.GetString(5);
 
-        Console.WriteLine($"CATEGORIA: {cName}\nTIPO DA CATEGORIA: {cType}\nid da transação: {tId}\ndescrição da transação: {tDescription}\nvalor da transação: {tAmount}\ndata da transação: {tDate}\n\n");
+        Console.WriteLine($"categoria: {cName}\ntipo da categoria: {cType}\n\nid da transação: {tId}\ndescrição da transação: {tDescription}\nvalor da transação: {tAmount}\ndata da transação: {tDate}\n\n\n");
         quantidadeTransacoes++;
+        idsValidos.Add(tId);
       }
+
       reader.Close();
+
       if (quantidadeTransacoes < 1)
       {
         System.Console.WriteLine("não existem transações para excluir.");
         return;
       }
 
-      System.Console.WriteLine("insira o id da transação que quer excluir: ");
+      System.Console.WriteLine("insira o id da transação que quer excluir: (0 para sair) ");
       bool validTransactionDeleteId = int.TryParse(Console.ReadLine(), out int tIdDelete);
 
-      while (!validTransactionDeleteId || tIdDelete < 0 || tIdDelete > quantidadeTransacoes)
+      if (tIdDelete == 0) return;
+
+      while (!validTransactionDeleteId || !idsValidos.Contains(tIdDelete))
       {
         System.Console.WriteLine("incorreto. insira novamente: ");
         validTransactionDeleteId = int.TryParse(Console.ReadLine(), out tIdDelete);
+
+        if (tIdDelete == 0) return;
       }
 
       command.CommandText = "DELETE FROM transactions WHERE id = $idInput";
       command.Parameters.AddWithValue("$idInput", tIdDelete);
       command.ExecuteNonQuery();
+      Console.ReadKey();
     }
 
     static void VerSaldoAtual(string connectionString)
     {
+      Console.Clear();
       Console.WriteLine("\nver saldo atual");
 
       double totalEntrada = 0;
@@ -307,6 +400,7 @@ namespace FinanceTracker
 
       double saldo = totalEntrada - totalSaida;
       System.Console.WriteLine($"saldo atual: R${saldo:F2}.");
+      Console.ReadKey();
     }
   }
 }
